[Test] Cleanup mailing tests
