+++ tests/Controller/CustomerLoginApiTest.php
@@ -1,0 +1,80 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Sylius\ShopApiPlugin\Controller;
+
+use Lakion\ApiTestCase\JsonApiTestCase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\Response;
+use Tests\Sylius\ShopApiPlugin\Controller\Utils\PurgeSpooledMessagesTrait;
+
+/**
+ * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
+ */
+final class CustomerLoginApiTest extends JsonApiTestCase
+{
+    use PurgeSpooledMessagesTrait;
+
+    /**
+     * @test
+     */
+    public function it_allows_to_log_customer_in(): void
+    {
+        $this->loadFixturesFromFile('customer.yml');
+
+        $data =
+<<<EOT
+        {
+            "_username": "oliver@queen.com",
+            "_password": "123pa\$\$word"
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/login_check', [], [], ['CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json'], $data);
+
+        $response = $this->client->getResponse();
+        $this->assertResponse($response, 'customer/customer_log_in_response', Response::HTTP_OK);
+    }
+
+    /**
+     * @test
+     */
+    public function it_requires_to_verify_email_address_for_newly_created_customers(): void
+    {
+        $data =
+<<<EOT
+        {
+            "firstName": "Vin",
+            "lastName": "Diesel",
+            "email": "vinny@fandf.com",
+            "user": {
+                "plainPassword" : {
+                    "first": "somepass",
+                    "second": "somepass"
+                }
+            }
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/register', [], [], ['CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json'], $data);
+
+        $data =
+<<<EOT
+        {
+            "_username": "vinny@fandf.com",
+            "_password": "somepass"
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/login_check', [], [], ['CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json'], $data);
+
+        $response = $this->client->getResponse();
+        $this->assertResponse($response, 'customer/bad_credentials_response', Response::HTTP_UNAUTHORIZED);
+    }
+
+    protected function getContainer(): ContainerInterface
+    {
+        return static::$sharedKernel->getContainer();
+    }
+}
