+++ tests/Controller/CustomerRegisterApiTest.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Sylius\ShopApiPlugin\Controller;
+
+use Lakion\ApiTestCase\JsonApiTestCase;
+use Sylius\Component\Core\Test\Services\EmailCheckerInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\Response;
+use Tests\Sylius\ShopApiPlugin\Controller\Utils\PurgeSpooledMessagesTrait;
+
+final class CustomerRegisterApiTest extends JsonApiTestCase
+{
+    use PurgeSpooledMessagesTrait;
+
+    /**
+     * @test
+     */
+    public function it_allows_to_register_in_shop(): void
+    {
+        $data =
+<<<EOT
+        {
+            "firstName": "Vin",
+            "lastName": "Diesel",
+            "email": "vinny@fandf.com",
+            "user": {
+                "plainPassword" : {
+                    "first": "somepass",
+                    "second": "somepass"
+                }
+            }
+        }
+EOT;
+
+        $this->client->request('POST', '/shop-api/register', [], [], ['CONTENT_TYPE' => 'application/json', 'ACCEPT' => 'application/json'], $data);
+
+        $response = $this->client->getResponse();
+        $this->assertResponseCode($response, Response::HTTP_CREATED);
+
+        /** @var EmailCheckerInterface $emailChecker */
+        $emailChecker = $this->get('sylius.behat.email_checker');
+
+        $this->assertTrue($emailChecker->hasRecipient('vinny@fandf.com'));
+    }
+
+    protected function getContainer(): ContainerInterface
+    {
+        return static::$sharedKernel->getContainer();
+    }
+}
